Add mineTimeEstimate()
mineTimeEstimate calculates expected time until next block successfully mined, including uncles. Note that this takes into account geth currently misreporting the hashrate by a factor of 1000. When that gets fixed, this will have to be updated (by pulling out the 1000* )
This assumes also that there are at least 2 uncles included every 25 blocks, on average. This does not actually change the prediction by much, but if uncles become particularly common, (say, an average of 1 per block) then the 1.04 would be changed to reflect this (e.g. to 2.0)