Add mko() for monitoring while mining.
mko() displays the coinbaseBalance() as well as the mineTimeEstimate(), and ALSO keeps track of the hashrate and the most recent block number.
If either of these are exactly the same as they were the last time mko() was called, it suggests that there might be a problem, so it gives the user an error warning them. Neither of these imply that there necessarily is a problem, but they might suggest that one might want to check on things.
The function is called mko because it relates to mining, so it starts with m, and the keys are right in a row on the way to the parens.